fix(server-mysql): address review findings on the MySQL backend
- add_records: replace INSERT IGNORE with INSERT ... ON DUPLICATE KEY
  UPDATE id = id, so only true duplicate-key rows are absorbed and
  genuine row errors (e.g. data-too-long) still abort the transaction,
  matching Postgres/SQLite ON CONFLICT DO NOTHING semantics
- store: widen data/cek/version to LONGTEXT; MySQL TEXT (64KB cap) would
  silently truncate large encrypted payloads, which INSERT IGNORE then
  committed as success (permanent, undecryptable corruption)
- store/store_idx_cache: pin tag to utf8mb4_bin for byte-exact matching
- users/sessions: BIGINT ids to match the i64 models and Postgres
- created_at: DATETIME instead of Y2038-limited TIMESTAMP
- DbSettings Debug: redact mysql:// URIs (redaction was Postgres-only)
- server-mysql crate: inherit workspace lints; fix explicit_iter_loop

+++ crates/atuin-server-mysql/migrations/20240108124831_create-sessions.sql
@@ -1,6 +1,6 @@
 create table sessions (
-  id int auto_increment primary key,
-	user_id integer,
+  id bigint auto_increment primary key,
+	user_id bigint,
 	token varchar(256) unique not null
 );
